linux/vboxsf: Another build fix (generic code now). bugref:9172
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@77881 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/VBox/VBoxGuestLibSharedFoldersInline.h
@@ -1404,7 +1404,7 @@ DECLINLINE(int) VbglR0SfHostReqReadLinkContigSimple(SHFLROOT idRoot, const char 
     if (cchPath < _64K - 1)
     {
         VBOXSFREADLINKREQ *pReq = (VBOXSFREADLINKREQ *)VbglR0PhysHeapAlloc(RT_UOFFSETOF(VBOXSFREADLINKREQ, StrPath.String)
-                                                                           + SHFLSTRING_HEADER_SIZE + cchPath);
+                                                                           + SHFLSTRING_HEADER_SIZE + (uint32_t)cchPath);
         if (pReq)
         {
             pReq->StrPath.u16Length = (uint16_t)cchPath;
